requirements.txt
Flask==2.2.3
pillow==9.3.0
pytesseract==0.3.9
flask-mail==0.9.1
gunicorn==20.1.0
python-dotenv==0.21.0
